+++ spec/dummy/app/models/post.rb
@@ -6,21 +6,5 @@
 class Post < ApplicationRecord
   include WhittakerTech::Oscar::Stateful
 
-  oscar_taxonomy(
-    initial: :draft,
-    states: {
-      draft: {},
-      published: {},
-      archived: {},
-      trashed: {},
-      purged: { destroyable: true, locked: true }
-    },
-    transitions: {
-      publish: { from: :draft, to: :published, past: :published },
-      archive: { from: :published, to: :archived, past: :archived },
-      trash: { from: %i[draft published archived], to: :trashed, past: :trashed },
-      restore: { from: :trashed, to: :draft, past: :draft },
-      purge: { from: :trashed, to: :purged, past: :purged }
-    }
-  )
+  oscar_taxonomy(base: :blog_post_visibility)
 end